Extensions: Make Python Integration work
