+++ src/App/Application.cpp
@@ -1138,6 +1138,7 @@ void Application::initTypes(void)
     App ::Extension                 ::init();
     App ::DocumentObjectExtension   ::init();
     App ::GroupExtension            ::init();
+    App ::GroupExtensionPython      ::init();
     App ::GeoFeatureGroupExtension  ::init();
     App ::OriginGroupExtension      ::init();
